requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.3
djangorestframework==3.15.2
fonttools==4.55.3
gunicorn==23.0.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
mysql-connector-python==9.1.0
mysqlclient==2.2.6
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
djangorestframework==3.15.2
scikit-learn==1.6.1
six==1.17.0
sqlparse==0.5.2
tzdata==2024.2
